Scope Plugin to All Non-Global Areas
The setup:upgrade command cannot complete when the before plugins execute
in the global scope. From some initial investigation; it would seem that,
during the execution of this command, Magento wants to (and expects)
indexers to be set to update on save but these plugins were preventing
that from happening - thus causing an exception.

Scoping these plugins to the different areas of the application does not
pose this problem - which is a solution I am okay with at the moment.

Happy for somebody to correct me and/or provide a better alternative
solution.